java: add `AnimateContainer`

+++ base/src/main/java/edu/hitsz/aircraft/HeroAircraft.java
@@ -49,10 +49,10 @@ public class HeroAircraft extends AbstractAircraft {
     @Override
     public LinkedList<BaseBullet> shoot() {
         LinkedList<BaseBullet> res = new LinkedList<>();
-        int x = this.getLocationX();
-        int y = this.getLocationY() + direction * 2;
-        int speedX = 0;
-        int speedY = this.getSpeedY() + direction * 5;
+        double x = this.getLocationX();
+        double y = this.getLocationY() + direction * 2;
+        double speedX = 0;
+        double speedY = this.getSpeedY() + direction * 5;
         for (int i = 0; i < shootNum; i++) {
             // 子弹发射位置相对飞机位置向前偏移
             // 多个子弹横向分散
